# docs/swagger.yaml
definitions:
  chum.AddReq:
    properties:
      friend_id:
        description: 朋友ID
        type: string
      permissions:
        description: 权限
        type: string
      source:
        description: 来源
        type: integer
      uuid:
        description: 本人ID
        type: string
    type: object
  chum.AddRes:
    properties:
      body:
        type: string
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  chum.ResChum:
    properties:
      body:
        properties:
          source:
            description: 搜索方式
            type: integer
          user:
            $ref: '#/definitions/user.Message'
            description: 用户信息
        type: object
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  getCode.GetPhoneCode:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  getCode.ResCode:
    properties:
      body:
        $ref: '#/definitions/getCode.GetPhoneCode'
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  login.User:
    properties:
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
  login.UserBody:
    properties:
      body:
        $ref: '#/definitions/login.User'
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  login.UserCode:
    properties:
      code:
        type: string
      phone:
        type: string
    required:
    - code
    - phone
    type: object
  login.UserName:
    properties:
      password:
        description: 用户密码
        type: string
      phone:
        description: 用户名
        type: string
    required:
    - password
    - phone
    type: object
  login.UserSignReps:
    properties:
      body:
        type: string
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  login.UserSignType:
    properties:
      code:
        description: 手机验证码
        type: string
      password:
        description: 用户密码
        type: string
      phone:
        description: 用户名
        type: string
    required:
    - code
    - password
    - phone
    type: object
  module.HttpErrs:
    properties:
      mgsText:
        description: 报错信息
        type: string
      msgCode:
        description: 状态码
        type: integer
    required:
    - mgsText
    - msgCode
    type: object
  resHome.KeysRes:
    properties:
      body:
        properties:
          private:
            description: 私有密钥
            type: string
          public:
            description: 共有密钥
            type: string
        type: object
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  resHome.MessageRes:
    properties:
      body:
        $ref: '#/definitions/user.Message'
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  resHome.MessageUpdateRes:
    properties:
      body:
        $ref: '#/definitions/user.Message'
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  uploadFiles.FileTest:
    properties:
      code:
        type: string
    type: object
  uploadFiles.FilesRes:
    properties:
      body:
        type: string
      mgsCode:
        description: 返回体状态码
        type: integer
      mgsText:
        description: 返回体信息
        type: string
    required:
    - mgsCode
    - mgsText
    type: object
  user.Message:
    properties:
      birthday:
        description: 生日
        type: string
      image:
        description: 头像
        type: string
      introduce:
        description: 介绍
        type: string
      name:
        description: 用户名称
        type: string
      phone:
        description: 用户手机
        type: string
      region:
        description: 地区
        type: string
      sex:
        description: 性别
        type: integer
      state:
        description: 状态 0 离线 1 在线
        type: integer
      user_id:
        description: 用户设置ID
        type: string
      uuid:
        description: 用户uuid
        type: string
    type: object
info:
  contact: {}
paths:
  /api/chum/add:
    post:
      consumes:
      - application/json
      operationId: AddUserFriendPaths
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 发送数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/chum.AddReq'
      produces:
      - application/json
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/chum.AddRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/module.HttpErrs'
      summary: 用户验证码登录
      tags:
      - Chum
  /api/chum/search/{phone}:
    get:
      consumes:
      - application/json
      operationId: SearchUserPaths
      parameters:
      - description: 朋友手机号
        in: path
        name: phone
        required: true
        type: string
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/chum.ResChum'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/module.HttpErrs'
      summary: 用户验证码登录
      tags:
      - Chum
  /api/home/key:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resHome.KeysRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/module.HttpErrs'
      summary: 获取数据加密
      tags:
      - Home
  /api/home/message:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/module.HttpErrs'
      summary: 首页接口
      tags:
      - Home
  /api/login/user/login:
    post:
      operationId: LoginInPasswordPaths
      parameters:
      - description: JSON数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/login.UserName'
      responses:
        "200":
          description: JSON数据
          schema:
            items:
              $ref: '#/definitions/login.UserBody'
            type: array
      summary: 用户登录操作
      tags:
      - Login
  /api/login/user/login_code:
    post:
      operationId: LoginsUserCode
      parameters:
      - description: JSON数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/login.UserCode'
      responses:
        "200":
          description: JSON数据
          schema:
            items:
              $ref: '#/definitions/login.UserBody'
            type: array
      summary: 用户登录操作
      tags:
      - Login
  /api/login/user/sign:
    post:
      operationId: LoginNewsUserPaths
      parameters:
      - description: JSON数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/login.UserSignType'
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/login.UserSignReps'
      summary: 用户注册操作
      tags:
      - Login
  /api/phone_code/user/{phone}:
    get:
      consumes:
      - application/json
      operationId: GetPathsCode
      parameters:
      - description: 用户手机号
        in: path
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/getCode.ResCode'
      summary: 用户验证码登录
      tags:
      - Code
  /api/upload/deleteImg:
    get:
      operationId: UploadImagesDelete
      parameters:
      - description: 图片名称
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/uploadFiles.FilesRes'
      summary: 删除图片操作
      tags:
      - Upload
  /api/upload/deleteVideo:
    get:
      operationId: UploadVideoDelete
      parameters:
      - description: 视频名称
        in: query
        name: video
        required: true
        type: string
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/uploadFiles.FilesRes'
      summary: 删除图片操作
      tags:
      - Upload
  /api/upload/images:
    post:
      operationId: UploadImages
      parameters:
      - description: file
        in: formData
        name: image
        required: true
        type: file
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/uploadFiles.FilesRes'
      summary: 上传图片操作
      tags:
      - Upload
  /api/upload/test:
    post:
      operationId: UploadTestPaths
      parameters:
      - description: 视频名称
        in: body
        name: video
        required: true
        schema:
          $ref: '#/definitions/uploadFiles.FileTest'
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/uploadFiles.FilesRes'
      summary: 删除图片操作
      tags:
      - Upload
  /api/upload/videos:
    post:
      operationId: UploadVideos
      parameters:
      - description: file
        in: formData
        name: videos
        required: true
        type: file
      responses:
        "200":
          description: JSON数据
          schema:
            $ref: '#/definitions/uploadFiles.FilesRes'
      summary: 上传视频操作
      tags:
      - Upload
  /api/user/user_message/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resHome.MessageRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/module.HttpErrs'
      summary: 获取个人信息
      tags:
      - User
  /api/user/user_message/update:
    post:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: 修改的数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.Message'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resHome.MessageUpdateRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/module.HttpErrs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/module.HttpErrs'
      summary: 修改个人信息
      tags:
      - User
swagger: "2.0"